projects/me/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
gunicorn==20.1.0
idna==3.2
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
